Solution to Input Two Floats and Print Their Product

+++ Input Two Floats and Print Their Product@@6788d5012db25efec946f93e/code.c
@@ -1,13 +1,13 @@
 #include <stdio.h>
 int main()
 {
-    int a,b,product;
+    float a,b,product;
 
-    scanf("%d %d",&a,&b);
+    scanf("%f %f",&a,&b);
 
     product= a*b;
 
-    printf("Product: %d",product);
+    printf("Product: %f",product);
 
     return 0;
 }
